Update configuration and data paths in txt2img model
This commit mainly corrects the paths of the stable-diffusion checkpoints and modifies the structure of the configuration file for better readability. It also fixes wrong file names in the test and training data sections and updates the associated classes in the "ldm/data/custom.py" file. In addition, unnecessary blank lines and comments have been addressed to improve the clarity of the code.